+++ CHANGELOG.md
@@ -10,6 +10,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 
 ##### Bug Fixes
 
+* Fix a crash in the analyzer when target deduplication is deactivated.
+  [Marius Rackwitz](https://github.com/mrackwitz)
+  [#4751](https://github.com/CocoaPods/CocoaPods/issues/4751)
+
 * Handle CoreData mapping models with recursive resource globs.
   [Eric Firestone](https://github.com/efirestone)
   [#4809](https://github.com/CocoaPods/CocoaPods/pull/4809)